refactor: move code, clean up, rename for clarity
