+++ workflows/flows/analyse_study_tasks/shared/copy_v6_pipeline_results.py
@@ -622,48 +622,3 @@ def copy_pipeline_results(
     logger.info(f"{pipeline_name} mover flowrun is {flowrun}")
 
     return True, []
-
-
-@task(
-    description="Copy Out-of-Production Assembly Results",
-    task_run_name="Copy out-of-production assembly results to {destination_root}",
-)
-def copy_out_of_production_analysis_results_to_destination_folder(
-    analyses: list[Analysis],
-    results_workspace: Path,
-    destination_root: str | Path,
-    timeout: int = 14400,
-) -> list[BatchCopyResult]:
-    """
-    FIXME: This function is very similar to copy_assembly_batch_results_to_destination_folder,
-    needs refactoring (ochkalova).
-
-    The source workspace is expected to have three pipeline subdirectories:
-    ``asa/``, ``virify/``, and ``map/``, each with per-assembly result folders.
-
-    :param analyses: List of analyses to copy results for
-    :param results_workspace: Source workspace with ``asa/``, ``virify/`, and ``map/`` subdirectories
-    :param destination_root: The root directory to copy results into
-    :param timeout: Timeout in seconds for each move operation (default: 4 hours)
-    :return: One copy result per ASA-completed analysis in the batch
-    """
-    logger = get_run_logger()
-
-    destination_root = Path(destination_root)
-
-    copy_results: list[BatchCopyResult] = []
-    for analysis in analyses:
-        copy_results.append(
-            copy_single_analysis_results(
-                analysis=analysis,
-                results_workspace=results_workspace,
-                destination_root=destination_root,
-                timeout=timeout,
-            )
-        )
-
-    logger.info(
-        f"Completed copying out-of-production results for {len(copy_results)} analyses to {destination_root}"
-    )
-
-    return copy_results
