+++ workflows/flows/analysis/assembly/flows/import_out_of_production_assembly_analysis_results.py
@@ -19,7 +19,8 @@ from workflows.ena_utils.ena_api_requests import (
     get_study_assemblies_from_ena,
 )
 from workflows.flows.analyse_study_tasks.shared.copy_v6_pipeline_results import (
-    copy_out_of_production_analysis_results_to_destination_folder,
+    BatchCopyResult,
+    copy_single_analysis_results,
     copy_v6_study_summaries,
 )
 from workflows.flows.analysis.assembly.flows.analysis_assembly_study import (
@@ -54,15 +55,15 @@ from workflows.models import Analysis, AssemblyAnalysisPipeline, Study
 
 
 @flow(
-    name="Ingest external assembly analysis results",
-    flow_run_name="Ingest assembly results: {results_dir}",
+    name="Import out-of-production assembly analysis results",
+    flow_run_name="Import out-of-production assembly analysis results: {results_dir}",
 )
-def external_assembly_analysis_ingestion(
+def import_out_of_production_assembly_analysis_results(
     results_dir: str,
     samplesheet_path: str,
 ):
     """
-    Ingest Assembly Analysis Pipeline results that were run out-of-production.
+    Import Assembly Analysis Pipeline results that were run out-of-production.
 
     This flow takes samplesheet, a path to out-of-production ASA/VIRify/MAP results and:
     1. Creates/fetches the required database models (study, assemblies and analyses)
@@ -84,7 +85,9 @@ def external_assembly_analysis_ingestion(
     if not samplesheet_path.is_file():
         raise FileNotFoundError(f"Samplesheet file does not exist: {samplesheet_path}")
 
-    logger.info(f"Starting external assembly ingestion from {results_dir}")
+    logger.info(
+        f"Starting import of out-of-production assembly analysis results from {results_dir}"
+    )
 
     # =========================================================================
     # STEP 1: Parse samplesheet structure and extract assembly accessions
@@ -156,7 +159,7 @@ def external_assembly_analysis_ingestion(
             logger,
             prompt_description=_(f"""\
                         **External Assembly Analysis Ingestion**
-                        Ingesting externally-run Assembly Analysis results for study {study_accession} \
+                        Ingesting out-of-production Assembly Analysis results for study {study_accession} \
                         produced by [Assembly Analysis Pipeline V6](https://github.com/EBI-Metagenomics/assembly-analysis-pipeline).
                         """),
         )
@@ -176,28 +179,28 @@ def external_assembly_analysis_ingestion(
     # =========================================================================
     logger.info("Step 5: Validating and importing pipeline results...")
 
-    logger.info("Setting ASA analysis states...")
+    logger.info("Setting ASA analysis states based on the end-of-execution reports...")
     set_asa_analysis_states(
         results_path / "asa",  # may as well be mgnify_study.results_dir / "asa"
         exported_analyses,
     )
 
     logger.info("Processing ASA results...")
-    process_external_pipeline_results(
+    process_out_of_production_analysis_results(
         exported_analyses,
         results_path,
         AssemblyAnalysisPipeline.ASA,
     )
 
     logger.info("Processing VIRify results...")
-    process_external_pipeline_results(
+    process_out_of_production_analysis_results(
         exported_analyses,
         results_path,
         AssemblyAnalysisPipeline.VIRIFY,
     )
 
     logger.info("Processing MAP results...")
-    process_external_pipeline_results(
+    process_out_of_production_analysis_results(
         exported_analyses,
         results_path,
         AssemblyAnalysisPipeline.MAP,
@@ -344,16 +347,19 @@ def _validate_results_structure(
 
 
 @flow(
-    flow_run_name="Import externally generated analysis results: {pipeline_type}",
+    flow_run_name="Import out-of-production generated analysis results: {pipeline_type}",
     retries=2,
     retry_delay_seconds=60,
 )
-def process_external_pipeline_results(
+def process_out_of_production_analysis_results(
     analyses: list[Analysis],
     results_path: Path,
     pipeline_type: AssemblyAnalysisPipeline,
 ) -> None:
-    """Import pipeline results without using batch abstraction."""
+    """
+    Import pipeline results without using batch abstraction.
+    # TODO: add docstring and description of the flow
+    """
     logger = get_run_logger()
 
     schema = None
@@ -410,19 +416,19 @@ def process_external_pipeline_results(
 
 
 @task
-def set_asa_analysis_states(assembly_current_outdir: Path, analyses: List[Analysis]):
+def set_asa_analysis_states(study_results_dir: Path, analyses: List[Analysis]):
     """
     This function processes the end-of-execution reports generated by the assembly analysis pipeline
     to determine the outcome of each assembly and updates their statuses in the database accordingly.
 
-    :param assembly_current_outdir: Path to the directory containing the pipeline output
+    :param study_results_dir: Path to the directory containing the pipeline output
     :param analyses: List of assembly analyses objects to process
     :raises ValueError: If the analysed_assemblies.csv file is missing
     """
     logger = get_run_logger()
 
-    qc_failed_csv = assembly_current_outdir / "qc_failed_assemblies.csv"
-    analysed_assemblies_csv = assembly_current_outdir / "analysed_assemblies.csv"
+    qc_failed_csv = study_results_dir / "qc_failed_assemblies.csv"
+    analysed_assemblies_csv = study_results_dir / "analysed_assemblies.csv"
     qc_failed_assemblies, analysed_assemblies = parse_asa_end_of_run_reports(
         qc_failed_csv, analysed_assemblies_csv, logger=logger
     )
@@ -494,3 +500,48 @@ def copy_out_of_production_assembly_analysis_results(
 
     study.external_results_dir = str(study_prefix)
     study.save(update_fields=["external_results_dir"])
+
+
+@task(
+    description="Copy Out-of-Production Assembly Results",
+    task_run_name="Copy out-of-production assembly results to {destination_root}",
+)
+def copy_out_of_production_analysis_results_to_destination_folder(
+    analyses: list[Analysis],
+    results_workspace: Path,
+    destination_root: str | Path,
+    timeout: int = 14400,
+) -> list[BatchCopyResult]:
+    """
+    FIXME: This function is very similar to copy_assembly_batch_results_to_destination_folder,
+    needs refactoring (ochkalova).
+
+    The source workspace is expected to have three pipeline subdirectories:
+    ``asa/``, ``virify/``, and ``map/``, each with per-assembly result folders.
+
+    :param analyses: List of analyses to copy results for
+    :param results_workspace: Source workspace with ``asa/``, ``virify/`, and ``map/`` subdirectories
+    :param destination_root: The root directory to copy results into
+    :param timeout: Timeout in seconds for each move operation (default: 4 hours)
+    :return: One copy result per ASA-completed analysis in the batch
+    """
+    logger = get_run_logger()
+
+    destination_root = Path(destination_root)
+
+    copy_results: list[BatchCopyResult] = []
+    for analysis in analyses:
+        copy_results.append(
+            copy_single_analysis_results(
+                analysis=analysis,
+                results_workspace=results_workspace,
+                destination_root=destination_root,
+                timeout=timeout,
+            )
+        )
+
+    logger.info(
+        f"Completed copying out-of-production results for {len(copy_results)} analyses to {destination_root}"
+    )
+
+    return copy_results
